requirements.txt
gunicorn
flask
tensorflow-cpu
datetime
scikit-learn
pandas
googletrans==3.1.0a0
